+++ jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   preset: "ts-jest/presets/js-with-babel-esm",
   testEnvironment: "node",
-  testMatch: ['**/jest/*.test.ts'],
+  testMatch: ['**/unit/**/*.test.ts'],
   transformIgnorePatterns: ['/node_modules']
 };
