+++ package.json
@@ -4,7 +4,9 @@
   "version": "0.1.12",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "tyoe": "module",
+  "engines": {
+    "node": ">=20.0.0"
+  }
   "files": [
     "dist"
   ],
@@ -20,8 +22,7 @@
     "build": "tsc",
     "format": "eslint --fix --quiet './src/**/*.ts*'",
     "lint": "eslint './src/**/*.ts*'",
-    "test": "env TS_NODE_PROJECT=tsconfig.test.json  NODE_OPTIONS=\"--loader ts-node/esm\" mocha --require ts-node/register test/unit/**/*.test.ts --exit",
-    "sample-test": "env TS_NODE_PROJECT=tsconfig.test.json  NODE_OPTIONS=\"--loader ts-node/esm\" mocha --require ts-node/register test/sample/**/*.test.ts",
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "mocha": "mocha-parallel-tests -r ts-node/register  --exit"
   },
   "devDependencies": {
@@ -56,7 +57,6 @@
     "@aws-sdk/client-s3": "^3.338.0",
     "@aws-sdk/credential-providers": "^3.338.0",
     "@aws-sdk/lib-storage": "^3.338.0",
-    "@tanker/file-reader": "^4.1.1",
     "axios": "^0.27.2",
     "crypto": "^1.0.1",
     "crypto-js": "^4.1.1",
