+++ src/vwbl/erc1155/VWBL.ts
@@ -19,7 +19,7 @@ import { ExtractMetadata, Metadata, PlainMetadata } from "../metadata";
 import {
   ConstructorProps,
   EncryptLogic,
-  EthersConstructorProps,
+  EthersConstructorProps, FileOrPath,
   GasSettings,
   ProgressSubscriber,
   StepStatus,
@@ -107,8 +107,8 @@ export class VWBLERC1155 extends VWBLBase {
     name: string,
     description: string,
     amount: number,
-    plainFile: File | File[],
-    thumbnailImage: File,
+    plainFile: FileOrPath | FileOrPath[],
+    thumbnailImage: FileOrPath,
     feeNumerator: number,
     encryptLogic: EncryptLogic = "base64",
     uploadEncryptedFileCallback?: UploadEncryptedFile,
@@ -148,13 +148,16 @@ export class VWBLERC1155 extends VWBLBase {
     const isRunningOnBrowser = typeof window !== "undefined";
     const encryptedDataUrls = await Promise.all(
       plainFileArray.map(async (file) => {
+        const plainFileBlob = file instanceof File ? file : new File([await fs.promises.readFile(file)],file);
+        const filePath = file instanceof File ? file.name : file;
+        const fileName: string = file instanceof File ? file.name : file.split("/").slice(-1)[0]; //ファイル名の取得だけのためにpathを使いたくなかった
         const encryptedContent =
           encryptLogic === "base64"
-            ? encryptString(await toBase64FromBlob(file), key)
+            ? encryptString(await toBase64FromBlob(plainFileBlob), key)
             : isRunningOnBrowser
-            ? await encryptFile(file, key)
-            : encryptStream(fs.createReadStream((file as any).path), key);
-        return await uploadEncryptedFunction(file.name, encryptedContent, uuid, awsConfig);
+            ? await encryptFile(plainFileBlob, key)
+            : encryptStream(fs.createReadStream(filePath), key);
+        return await uploadEncryptedFunction(fileName, encryptedContent, uuid, awsConfig);
       })
     );
     const thumbnailImageUrl = await uploadThumbnailFunction(thumbnailImage, uuid, awsConfig);
@@ -218,8 +221,8 @@ export class VWBLERC1155 extends VWBLBase {
     name: string,
     description: string,
     amount: number,
-    plainFile: File | File[],
-    thumbnailImage: File,
+    plainFile: FileOrPath | FileOrPath[],
+    thumbnailImage: FileOrPath,
     feeNumerator: number,
     encryptLogic: EncryptLogic = "base64",
     subscriber?: ProgressSubscriber,
@@ -242,9 +245,11 @@ export class VWBLERC1155 extends VWBLBase {
     console.log("upload data");
     const encryptedDataUrls = await Promise.all(
       plainFileArray.map(async (file) => {
+        const plainFileBlob = file instanceof File ? file : new File([await fs.promises.readFile(file)],file);
+        const filePath = file instanceof File ? file.name : file;
+        const fileName: string = file instanceof File ? file.name : file.split("/").slice(-1)[0]; //ファイル名の取得だけのためにpathを使いたくなかった
         const encryptedContent =
-          encryptLogic === "base64" ? encryptString(await toBase64FromBlob(file), key) : await encryptFile(file, key);
-        console.log(typeof encryptedContent);
+          encryptLogic === "base64" ? encryptString(await toBase64FromBlob(plainFileBlob), key) : await encryptFile(plainFileBlob, key);
         return await this.uploadToIpfs?.uploadEncryptedFile(encryptedContent);
       })
     );
