+++ test/unit/vwbl/VWBL.test.ts
@@ -66,7 +66,7 @@ describe("VWBL with web3.js", () => {
   const uploadFileStub = sinon.stub(testFunctions, "uploadThumbnail").returns(Promise.resolve("https://example.com"));
   const uploadMetadataStub = sinon.stub(testFunctions, "uploadMetadata");
 
-  before(async () => {
+  beforeAll(async () => {
     await vwbl.sign();
   });
 
@@ -75,8 +75,8 @@ describe("VWBL with web3.js", () => {
     const tokenId = await vwbl.managedCreateToken(
       "test token",
       "test",
-      "asset/thumbnail.png",
-      "asset/plain.png",
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
       10,
       "base64",
       testFunctions.uploadEncryptedFile,
@@ -101,8 +101,8 @@ describe("VWBL with web3.js", () => {
     const tokenId = await vwbl.managedCreateToken(
       "test token",
       "test",
-      "asset/thumbnail.png",
-      "asset/plain.png",
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
       10,
       "base64",
       testFunctions.uploadEncryptedFile,
@@ -129,8 +129,8 @@ describe("VWBL with web3.js", () => {
     const tokenId = await vwbl.managedCreateToken(
       "test token",
       "test",
-      "asset/thumbnail.png",
-      "asset/plain.png",
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
       10,
       "base64",
       testFunctions.uploadEncryptedFile,
@@ -150,125 +150,125 @@ describe("VWBL with web3.js", () => {
   });
 });
 
-// describe("VWBLERC1155 with web3.js", () => {
-//   const vwblProtocolStub = {
-//     mintToken: sinon.stub(VWBLERC1155Contract.prototype, "mintToken"),
-//   };
-//
-//   const vwbl = new VWBLERC1155({
-//     ipfsNftStorageKey: "set nftstorage api key",
-//     awsConfig: undefined,
-//     contractAddress: "0x2c7e967093d7fe0eeb5440bf49e5D148417B0412",
-//     manageKeyType: ManageKeyType.VWBL_NETWORK_SERVER,
-//     uploadContentType: UploadContentType.CUSTOM,
-//     uploadMetadataType: UploadMetadataType.CUSTOM,
-//     vwblNetworkUrl: "http://example.com",
-//     web3,
-//   });
-//
-//   const testFunctions = {
-//     uploadEncryptedFile: () => {
-//       return Promise.resolve("https://example.com");
-//     },
-//     uploadThumbnail: () => {
-//       return Promise.resolve("https://example.com");
-//     },
-//     // eslint-disable-next-line @typescript-eslint/no-empty-function
-//     uploadMetadata: async () => {},
-//   };
-//   const uploadEncryptedFileStub = sinon
-//   .stub(testFunctions, "uploadEncryptedFile")
-//   .returns(Promise.resolve("https://example.com"));
-//   const uploadFileStub = sinon.stub(testFunctions, "uploadThumbnail").returns(Promise.resolve("https://example.com"));
-//   const uploadMetadataStub = sinon.stub(testFunctions, "uploadMetadata");
-//
-//   before(async () => {
-//     await vwbl.sign();
-//   });
-//
-//   it("mint erc1155 token without gas settings", async () => {
-//     vwblProtocolStub.mintToken.returns(Promise.resolve(1));
-//
-//     const tokenId = await vwbl.managedCreateToken(
-//       "test token",
-//       "test",
-//       100,
-//       "asset/thumbnail.png",
-//       "asset/plain.png",
-//       10,
-//       "base64",
-//       testFunctions.uploadEncryptedFile,
-//       testFunctions.uploadThumbnail,
-//       testFunctions.uploadMetadata
-//     );
-//
-//     expect(vwblProtocolStub.mintToken.callCount).equal(1);
-//     expect(vwblProtocolStub.mintToken.getCall(0).args[4]).equal(undefined);
-//     expect(vwblApiStub.setKey.callCount).equal(4);
-//     expect(uploadEncryptedFileStub.callCount).equal(1);
-//     expect(uploadFileStub.callCount).equal(1);
-//     expect(uploadMetadataStub.callCount).equal(1);
-//     expect(tokenId).equal(1);
-//   });
-//
-//   it("mint erc1155 token with maxPriorityFee and maxFee", async () => {
-//     vwblProtocolStub.mintToken.returns(Promise.resolve(2));
-//     const testSubscriber = {
-//       kickStep: () => {}
-//     }
-//     const tokenId = await vwbl.managedCreateToken(
-//       "test token",
-//       "test",
-//       100,
-//       "asset/thumbnail.png",
-//       "asset/plain.png",
-//       10,
-//       "base64",
-//       testFunctions.uploadEncryptedFile,
-//       testFunctions.uploadThumbnail,
-//       testFunctions.uploadMetadata,
-//       testSubscriber,
-//       {maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000}
-//     );
-//
-//     expect(vwblProtocolStub.mintToken.callCount).equal(2);
-//     expect(vwblProtocolStub.mintToken.getCall(1).args[4]).deep.equal({maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000});
-//     expect(vwblApiStub.setKey.callCount).equal(5);
-//     expect(uploadEncryptedFileStub.callCount).equal(2);
-//     expect(uploadFileStub.callCount).equal(2);
-//     expect(uploadMetadataStub.callCount).equal(2);
-//     expect(tokenId).equal(2);
-//   });
-//
-//   it("mint erc1155 token with gasPrice", async () => {
-//     vwblProtocolStub.mintToken.returns(Promise.resolve(3));
-//     const testSubscriber = {
-//       kickStep: () => {}
-//     }
-//     const tokenId = await vwbl.managedCreateToken(
-//       "test token",
-//       "test",
-//       100,
-//       "asset/thumbnail.png",
-//       "asset/plain.png",
-//       10,
-//       "base64",
-//       testFunctions.uploadEncryptedFile,
-//       testFunctions.uploadThumbnail,
-//       testFunctions.uploadMetadata,
-//       testSubscriber,
-//       {gasPrice: 1000}
-//     );
-//
-//     expect(vwblProtocolStub.mintToken.callCount).equal(3);
-//     expect(vwblProtocolStub.mintToken.getCall(2).args[4]).deep.equal({gasPrice: 1000});
-//     expect(vwblApiStub.setKey.callCount).equal(6);
-//     expect(uploadEncryptedFileStub.callCount).equal(3);
-//     expect(uploadFileStub.callCount).equal(3);
-//     expect(uploadMetadataStub.callCount).equal(3);
-//     expect(tokenId).equal(3);
-//   });
-// });
+describe("VWBLERC1155 with web3.js", () => {
+  const vwblProtocolStub = {
+    mintToken: sinon.stub(VWBLERC1155Contract.prototype, "mintToken"),
+  };
+
+  const vwbl = new VWBLERC1155({
+    ipfsNftStorageKey: "set nftstorage api key",
+    awsConfig: undefined,
+    contractAddress: "0x2c7e967093d7fe0eeb5440bf49e5D148417B0412",
+    manageKeyType: ManageKeyType.VWBL_NETWORK_SERVER,
+    uploadContentType: UploadContentType.CUSTOM,
+    uploadMetadataType: UploadMetadataType.CUSTOM,
+    vwblNetworkUrl: "http://example.com",
+    web3,
+  });
+
+  const testFunctions = {
+    uploadEncryptedFile: () => {
+      return Promise.resolve("https://example.com");
+    },
+    uploadThumbnail: () => {
+      return Promise.resolve("https://example.com");
+    },
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    uploadMetadata: async () => {},
+  };
+  const uploadEncryptedFileStub = sinon
+  .stub(testFunctions, "uploadEncryptedFile")
+  .returns(Promise.resolve("https://example.com"));
+  const uploadFileStub = sinon.stub(testFunctions, "uploadThumbnail").returns(Promise.resolve("https://example.com"));
+  const uploadMetadataStub = sinon.stub(testFunctions, "uploadMetadata");
+
+  beforeAll(async () => {
+    await vwbl.sign();
+  });
+
+  it("mint erc1155 token without gas settings", async () => {
+    vwblProtocolStub.mintToken.returns(Promise.resolve(1));
+
+    const tokenId = await vwbl.managedCreateToken(
+      "test token",
+      "test",
+      100,
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
+      10,
+      "base64",
+      testFunctions.uploadEncryptedFile,
+      testFunctions.uploadThumbnail,
+      testFunctions.uploadMetadata
+    );
+
+    expect(vwblProtocolStub.mintToken.callCount).equal(1);
+    expect(vwblProtocolStub.mintToken.getCall(0).args[4]).equal(undefined);
+    expect(vwblApiStub.setKey.callCount).equal(4);
+    expect(uploadEncryptedFileStub.callCount).equal(1);
+    expect(uploadFileStub.callCount).equal(1);
+    expect(uploadMetadataStub.callCount).equal(1);
+    expect(tokenId).equal(1);
+  });
+
+  it("mint erc1155 token with maxPriorityFee and maxFee", async () => {
+    vwblProtocolStub.mintToken.returns(Promise.resolve(2));
+    const testSubscriber = {
+      kickStep: () => {}
+    }
+    const tokenId = await vwbl.managedCreateToken(
+      "test token",
+      "test",
+      100,
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
+      10,
+      "base64",
+      testFunctions.uploadEncryptedFile,
+      testFunctions.uploadThumbnail,
+      testFunctions.uploadMetadata,
+      testSubscriber,
+      {maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000}
+    );
+
+    expect(vwblProtocolStub.mintToken.callCount).equal(2);
+    expect(vwblProtocolStub.mintToken.getCall(1).args[4]).deep.equal({maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000});
+    expect(vwblApiStub.setKey.callCount).equal(5);
+    expect(uploadEncryptedFileStub.callCount).equal(2);
+    expect(uploadFileStub.callCount).equal(2);
+    expect(uploadMetadataStub.callCount).equal(2);
+    expect(tokenId).equal(2);
+  });
+
+  it("mint erc1155 token with gasPrice", async () => {
+    vwblProtocolStub.mintToken.returns(Promise.resolve(3));
+    const testSubscriber = {
+      kickStep: () => {}
+    }
+    const tokenId = await vwbl.managedCreateToken(
+      "test token",
+      "test",
+      100,
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
+      10,
+      "base64",
+      testFunctions.uploadEncryptedFile,
+      testFunctions.uploadThumbnail,
+      testFunctions.uploadMetadata,
+      testSubscriber,
+      {gasPrice: 1000}
+    );
+
+    expect(vwblProtocolStub.mintToken.callCount).equal(3);
+    expect(vwblProtocolStub.mintToken.getCall(2).args[4]).deep.equal({gasPrice: 1000});
+    expect(vwblApiStub.setKey.callCount).equal(6);
+    expect(uploadEncryptedFileStub.callCount).equal(3);
+    expect(uploadFileStub.callCount).equal(3);
+    expect(uploadMetadataStub.callCount).equal(3);
+    expect(tokenId).equal(3);
+  });
+});
 
 describe("VWBL with ethers.js", () => {
   const vwblProtocolStub = {
@@ -303,7 +303,7 @@ describe("VWBL with ethers.js", () => {
   const uploadFileStub = sinon.stub(testFunctions, "uploadThumbnail").returns(Promise.resolve("https://example.com"));
   const uploadMetadataStub = sinon.stub(testFunctions, "uploadMetadata");
 
-  before(async () => {
+  beforeAll(async () => {
     await vwbl.sign();
   });
 
@@ -312,8 +312,8 @@ describe("VWBL with ethers.js", () => {
     const tokenId = await vwbl.managedCreateToken(
       "test token",
       "test",
-      "asset/thumbnail.png",
-      "asset/plain.png",
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
       10,
       "base64",
       testFunctions.uploadEncryptedFile,
@@ -338,8 +338,8 @@ describe("VWBL with ethers.js", () => {
     const tokenId = await vwbl.managedCreateToken(
       "test token",
       "test",
-      "asset/thumbnail.png",
-      "asset/plain.png",
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
       10,
       "base64",
       testFunctions.uploadEncryptedFile,
@@ -366,8 +366,8 @@ describe("VWBL with ethers.js", () => {
     const tokenId = await vwbl.managedCreateToken(
       "test token",
       "test",
-      "asset/thumbnail.png",
-      "asset/plain.png",
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
       10,
       "base64",
       testFunctions.uploadEncryptedFile,
@@ -387,123 +387,123 @@ describe("VWBL with ethers.js", () => {
   });
 });
 
-// describe("VWBLERC1155 with ethers.js", () => {
-//   const vwblProtocolStub = {
-//     mintToken: sinon.stub( VWBLERC1155EthersContract.prototype, "mintToken"),
-//   };
-//
-//   const vwbl = new VWBLERC1155({
-//     ipfsNftStorageKey: "set nftstorage api key",
-//     awsConfig: undefined,
-//     contractAddress: "0x2c7e967093d7fe0eeb5440bf49e5D148417B0412",
-//     manageKeyType: ManageKeyType.VWBL_NETWORK_SERVER,
-//     uploadContentType: UploadContentType.CUSTOM,
-//     uploadMetadataType: UploadMetadataType.CUSTOM,
-//     vwblNetworkUrl: "http://example.com",
-//     ethersProvider: ethProvider,
-//     ethersSigner: ethSigner,
-//   });
-//
-//   const testFunctions = {
-//     uploadEncryptedFile: () => {
-//       return Promise.resolve("https://example.com");
-//     },
-//     uploadThumbnail: () => {
-//       return Promise.resolve("https://example.com");
-//     },
-//     // eslint-disable-next-line @typescript-eslint/no-empty-function
-//     uploadMetadata: async () => {},
-//   };
-//   const uploadEncryptedFileStub = sinon
-//   .stub(testFunctions, "uploadEncryptedFile")
-//   .returns(Promise.resolve("https://example.com"));
-//   const uploadFileStub = sinon.stub(testFunctions, "uploadThumbnail").returns(Promise.resolve("https://example.com"));
-//   const uploadMetadataStub = sinon.stub(testFunctions, "uploadMetadata");
-//
-//   before(async () => {
-//     await vwbl.sign();
-//   });
-//
-//   it("mint erc1155 token without gas settings", async () => {
-//     vwblProtocolStub.mintToken.returns(Promise.resolve(1));
-//
-//     const tokenId = await vwbl.managedCreateToken(
-//       "test token",
-//       "test",
-//       100,
-//       "asset/thumbnail.png",
-//       "asset/plain.png",
-//       10,
-//       "base64",
-//       testFunctions.uploadEncryptedFile,
-//       testFunctions.uploadThumbnail,
-//       testFunctions.uploadMetadata
-//     );
-//
-//     expect(vwblProtocolStub.mintToken.callCount).equal(1);
-//     expect(vwblProtocolStub.mintToken.getCall(0).args[4]).equal(undefined);
-//     expect(vwblApiStub.setKey.callCount).equal(10);
-//     expect(uploadEncryptedFileStub.callCount).equal(1);
-//     expect(uploadFileStub.callCount).equal(1);
-//     expect(uploadMetadataStub.callCount).equal(1);
-//     expect(tokenId).equal(1);
-//   });
-//
-//   it("mint erc1155 token with maxPriorityFee and maxFee", async () => {
-//     vwblProtocolStub.mintToken.returns(Promise.resolve(2));
-//     const testSubscriber = {
-//       kickStep: () => {}
-//     }
-//     const tokenId = await vwbl.managedCreateToken(
-//       "test token",
-//       "test",
-//       100,
-//       "asset/thumbnail.png",
-//       "asset/plain.png",
-//       10,
-//       "base64",
-//       testFunctions.uploadEncryptedFile,
-//       testFunctions.uploadThumbnail,
-//       testFunctions.uploadMetadata,
-//       testSubscriber,
-//       {maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000}
-//     );
-//
-//     expect(vwblProtocolStub.mintToken.callCount).equal(2);
-//     expect(vwblProtocolStub.mintToken.getCall(1).args[4]).deep.equal({maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000});
-//     expect(vwblApiStub.setKey.callCount).equal(11);
-//     expect(uploadEncryptedFileStub.callCount).equal(2);
-//     expect(uploadFileStub.callCount).equal(2);
-//     expect(uploadMetadataStub.callCount).equal(2);
-//     expect(tokenId).equal(2);
-//   });
-//
-//   it("mint erc1155 token with gasPrice", async () => {
-//     vwblProtocolStub.mintToken.returns(Promise.resolve(3));
-//     const testSubscriber = {
-//       kickStep: () => {}
-//     }
-//     const tokenId = await vwbl.managedCreateToken(
-//       "test token",
-//       "test",
-//       100,
-//       "asset/thumbnail.png",
-//       "asset/plain.png",
-//       10,
-//       "base64",
-//       testFunctions.uploadEncryptedFile,
-//       testFunctions.uploadThumbnail,
-//       testFunctions.uploadMetadata,
-//       testSubscriber,
-//       {gasPrice: 1000}
-//     );
-//
-//     expect(vwblProtocolStub.mintToken.callCount).equal(3);
-//     expect(vwblProtocolStub.mintToken.getCall(2).args[4]).deep.equal({gasPrice: 1000});
-//     expect(vwblApiStub.setKey.callCount).equal(12);
-//     expect(uploadEncryptedFileStub.callCount).equal(3);
-//     expect(uploadFileStub.callCount).equal(3);
-//     expect(uploadMetadataStub.callCount).equal(3);
-//     expect(tokenId).equal(3);
-//   });
-// });
+describe("VWBLERC1155 with ethers.js", () => {
+  const vwblProtocolStub = {
+    mintToken: sinon.stub( VWBLERC1155EthersContract.prototype, "mintToken"),
+  };
+
+  const vwbl = new VWBLERC1155({
+    ipfsNftStorageKey: "set nftstorage api key",
+    awsConfig: undefined,
+    contractAddress: "0x2c7e967093d7fe0eeb5440bf49e5D148417B0412",
+    manageKeyType: ManageKeyType.VWBL_NETWORK_SERVER,
+    uploadContentType: UploadContentType.CUSTOM,
+    uploadMetadataType: UploadMetadataType.CUSTOM,
+    vwblNetworkUrl: "http://example.com",
+    ethersProvider: ethProvider,
+    ethersSigner: ethSigner,
+  });
+
+  const testFunctions = {
+    uploadEncryptedFile: () => {
+      return Promise.resolve("https://example.com");
+    },
+    uploadThumbnail: () => {
+      return Promise.resolve("https://example.com");
+    },
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    uploadMetadata: async () => {},
+  };
+  const uploadEncryptedFileStub = sinon
+  .stub(testFunctions, "uploadEncryptedFile")
+  .returns(Promise.resolve("https://example.com"));
+  const uploadFileStub = sinon.stub(testFunctions, "uploadThumbnail").returns(Promise.resolve("https://example.com"));
+  const uploadMetadataStub = sinon.stub(testFunctions, "uploadMetadata");
+
+  beforeAll(async () => {
+    await vwbl.sign();
+  });
+
+  it("mint erc1155 token without gas settings", async () => {
+    vwblProtocolStub.mintToken.returns(Promise.resolve(1));
+
+    const tokenId = await vwbl.managedCreateToken(
+      "test token",
+      "test",
+      100,
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
+      10,
+      "base64",
+      testFunctions.uploadEncryptedFile,
+      testFunctions.uploadThumbnail,
+      testFunctions.uploadMetadata
+    );
+
+    expect(vwblProtocolStub.mintToken.callCount).equal(1);
+    expect(vwblProtocolStub.mintToken.getCall(0).args[4]).equal(undefined);
+    expect(vwblApiStub.setKey.callCount).equal(10);
+    expect(uploadEncryptedFileStub.callCount).equal(1);
+    expect(uploadFileStub.callCount).equal(1);
+    expect(uploadMetadataStub.callCount).equal(1);
+    expect(tokenId).equal(1);
+  });
+
+  it("mint erc1155 token with maxPriorityFee and maxFee", async () => {
+    vwblProtocolStub.mintToken.returns(Promise.resolve(2));
+    const testSubscriber = {
+      kickStep: () => {}
+    }
+    const tokenId = await vwbl.managedCreateToken(
+      "test token",
+      "test",
+      100,
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
+      10,
+      "base64",
+      testFunctions.uploadEncryptedFile,
+      testFunctions.uploadThumbnail,
+      testFunctions.uploadMetadata,
+      testSubscriber,
+      {maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000}
+    );
+
+    expect(vwblProtocolStub.mintToken.callCount).equal(2);
+    expect(vwblProtocolStub.mintToken.getCall(1).args[4]).deep.equal({maxPriorityFeePerGas: 40000000000, maxFeePerGas: 41000000000});
+    expect(vwblApiStub.setKey.callCount).equal(11);
+    expect(uploadEncryptedFileStub.callCount).equal(2);
+    expect(uploadFileStub.callCount).equal(2);
+    expect(uploadMetadataStub.callCount).equal(2);
+    expect(tokenId).equal(2);
+  });
+
+  it("mint erc1155 token with gasPrice", async () => {
+    vwblProtocolStub.mintToken.returns(Promise.resolve(3));
+    const testSubscriber = {
+      kickStep: () => {}
+    }
+    const tokenId = await vwbl.managedCreateToken(
+      "test token",
+      "test",
+      100,
+      "test/asset/thumbnail.png",
+      "test/asset/plain.png",
+      10,
+      "base64",
+      testFunctions.uploadEncryptedFile,
+      testFunctions.uploadThumbnail,
+      testFunctions.uploadMetadata,
+      testSubscriber,
+      {gasPrice: 1000}
+    );
+
+    expect(vwblProtocolStub.mintToken.callCount).equal(3);
+    expect(vwblProtocolStub.mintToken.getCall(2).args[4]).deep.equal({gasPrice: 1000});
+    expect(vwblApiStub.setKey.callCount).equal(12);
+    expect(uploadEncryptedFileStub.callCount).equal(3);
+    expect(uploadFileStub.callCount).equal(3);
+    expect(uploadMetadataStub.callCount).equal(3);
+    expect(tokenId).equal(3);
+  });
+});
